Jwt filter working
-- On login will get the token
--  working on otp signup

+++ src/main/java/com/maaztausif/khallikarao/repository/AuthRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 @Repository
-public interface AuthRepo extends JpaRepository<User,Integer> {
+public interface AuthRepo extends JpaRepository<User,Long> {
     Optional<User> findByEmail(String email);
 }
